refactor: 관심멤버 페이지 서버 컴포넌트를 components디렉토리로 이동
