+++ src/app/mypage/hubprofile/page.tsx
@@ -1,6 +1,7 @@
 import HubProfile from "@/components/MyPage/HubInfo/HubProfile";
 
+const HubProfilePage = () => {
+  return <HubProfile />;
+};
 
-export default function HubProfilePage() {
-  return <HubProfile />; 
-}
+export default HubProfilePage;
