+++ src/app/mypage/myposts/page.tsx
@@ -1,11 +1,7 @@
 import PostsTap from "@/components/MyPage/MyPosts/PostsTap";
 
 const MyPostsPage = () => {
-  return (
-    <div>
-      <PostsTap />
-    </div>
-  );
+  return <PostsTap />;
 };
 
 export default MyPostsPage;
